% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathgenscrape-package.R
\docType{package}
\name{mathgenscrape-package}
\alias{mathgenscrape}
\alias{mathgenscrape-package}
\title{mathgenscrape: Scrape the Mathematics Genealogy Project}
\description{
Look up mathematicians by name (\code{\link[=mathgen_lookup]{mathgen_lookup()}},
\code{\link[=mathgen_advanced_search]{mathgen_advanced_search()}}) and recursively trace their advisor or
student lineage into an edge list (\code{\link[=get_advisors]{get_advisors()}}, \code{\link[=get_students]{get_students()}},
\code{\link[=mathgen_traverse]{mathgen_traverse()}}) suitable for graph analysis.
}
\author{
\strong{Maintainer}: Ben Bolker \email{bbolker@gmail.com}

Authors:
\itemize{
  \item Ben Bolker \email{bbolker@gmail.com}
}

}
\keyword{internal}
